requirements.txt
asgiref==3.5.2
Django==3.0
pytz==2022.4
sqlparse==0.4.3
gunicorn==20.0.4
psycopg2
requests==2.28.1
django-environ==0.9.0
djangorestframework==3.13.1
django-cors-headers==3.11.0